Adds changes after review of PR #150.
