+++ cortex/built_ins/models/vae.py
@@ -1,12 +1,11 @@
 '''Simple Variational Autoencoder model.
 '''
 
-from cortex.plugins import ModelPlugin
+from cortex.plugins import ModelPlugin, register_plugin
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch.autograd import Variable
-from cortex.main import run
 from cortex.built_ins.models.utils import update_encoder_args, update_decoder_args, ms_ssim
 
 __author__ = 'R Devon Hjelm and Samuel Lavoie'
@@ -39,8 +38,8 @@ class VAENetwork(nn.Module):
 
     def reparametrize(self, mu, std):
         if self.training:
-            esp = Variable(std.data.new(std.size()).normal_(),
-                           requires_grad=False).cuda()
+            esp = Variable(
+                std.data.new(std.size()).normal_(), requires_grad=False).cuda()
             return mu + std * esp
         else:
             return mu
@@ -59,7 +58,9 @@ class ImageEncoder(ModelPlugin):
 
     '''
 
-    def build(self, encoder_type: str = 'convnet', dim_out: int = None,
+    def build(self,
+              encoder_type: str = 'convnet',
+              dim_out: int = None,
               encoder_args=dict(fully_connected_layers=1028)):
         '''
 
@@ -70,9 +71,8 @@ class ImageEncoder(ModelPlugin):
 
         '''
         x_shape = self.get_dims('x', 'y', 'c')
-        Encoder, encoder_args = update_encoder_args(x_shape,
-                                                    model_type=encoder_type,
-                                                    encoder_args=encoder_args)
+        Encoder, encoder_args = update_encoder_args(
+            x_shape, model_type=encoder_type, encoder_args=encoder_args)
         encoder = Encoder(x_shape, dim_out=dim_out, **encoder_args)
         self.nets.encoder = encoder
 
@@ -91,7 +91,9 @@ class ImageDecoder(ModelPlugin):
 
     '''
 
-    def build(self, decoder_type: str = 'convnet', dim_in: int = 64,
+    def build(self,
+              decoder_type: str = 'convnet',
+              dim_in: int = 64,
               decoder_args=dict(output_nonlinearity='tanh')):
         '''
 
@@ -102,9 +104,8 @@ class ImageDecoder(ModelPlugin):
 
         '''
         x_shape = self.get_dims('x', 'y', 'c')
-        Decoder, decoder_args = update_decoder_args(x_shape,
-                                                    model_type=decoder_type,
-                                                    decoder_args=decoder_args)
+        Decoder, decoder_args = update_decoder_args(
+            x_shape, model_type=decoder_type, decoder_args=decoder_args)
         decoder = Decoder(x_shape, dim_in=dim_in, **decoder_args)
         decoder = Decoder(x_shape, dim_in=dim_in, **decoder_args)
         self.nets.decoder = decoder
@@ -143,11 +144,10 @@ class VAE(ModelPlugin):
     plugin_name = 'VAE'
 
     defaults = dict(
-        data=dict(batch_size=dict(train=64, test=640),
-                  inputs=dict(inputs='images')),
+        data=dict(
+            batch_size=dict(train=64, test=640), inputs=dict(inputs='images')),
         optimizer=dict(optimizer='Adam', learning_rate=1e-4),
-        train=dict(save_on_lowest='losses.vae')
-    )
+        train=dict(save_on_lowest='losses.vae'))
 
     def __init__(self):
         super().__init__()
@@ -172,8 +172,7 @@ class VAE(ModelPlugin):
         self.add_noise('Z', dist='normal', size=dim_z)
         encoder = self.nets.encoder
         decoder = self.nets.decoder
-        vae = VAENetwork(encoder, decoder, dim_out=dim_encoder_out,
-                         dim_z=dim_z)
+        vae = VAENetwork(encoder, decoder, dim_out=dim_encoder_out, dim_z=dim_z)
         self.nets.vae = vae
 
     def routine(self, inputs, targets, Z, vae_criterion=F.mse_loss,
@@ -191,8 +190,8 @@ class VAE(ModelPlugin):
 
         r_loss = vae_criterion(
             outputs, inputs, size_average=False) / inputs.size(0)
-        kl = (0.5 * (vae.std ** 2 + vae.mu ** 2 - 2. *
-                     torch.log(vae.std) - 1.).sum(1).mean())
+        kl = (0.5 * (vae.std**2 + vae.mu**2 - 2. * torch.log(vae.std) -
+                     1.).sum(1).mean())
 
         msssim = ms_ssim(inputs, outputs)
 
@@ -210,6 +209,4 @@ class VAE(ModelPlugin):
         self.decoder.visualize(vae.mu.data)
 
 
-if __name__ == '__main__':
-    classifier = VAE()
-    run(model=classifier)
+register_plugin(VAE)
